Add customizable InjectTemplateListener
The template listener checks for modules with more than a first
top level namespace indicating the module. This holds for cases
in the Soflomo family like Soflomo\Blog and Soflomo\Events.

The template listener resolves to a template name based on a
configurable number of top level namespaces. By default, both two
modules would resolve to a top level directory "soflomo" while with
a level of n=2 the directories become "soflomo-blog" and "soflomo-
events".

The listener is configurable to check the first (default Zend
behaviour), two (default behaviour of this listener) or more name-
spaces.

Usage example:

```
$listener = new Soflomo\Common\View\InjectTemplateListener;

$em = $application->getEventManager()->getSharedEventmanager();
$em->attach('My\Specific\Controller\FooController', MvcEvent::EVENT_DISPATCH,
  array($listener, 'injectTemplate'), -80);
```

Now when the My\Specific\Controller\FooController::SomeAction() is
dispatched, the view script is resolved to:

    my-specific/foo/some.phtml